[tf.data] Replacing `parallel_interleave` with `interleave` in the implementation of `TFRecordDataset`, `make_csv_dataset` and `make_batched_features_dataset`.
PiperOrigin-RevId: 237461383
