+++ tensorflow/python/data/ops/readers.py
@@ -26,7 +26,6 @@ from tensorflow.python.framework import ops
 from tensorflow.python.framework import tensor_shape
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import gen_dataset_ops
-from tensorflow.python.ops import gen_experimental_dataset_ops as ged_ops
 from tensorflow.python.util.tf_export import tf_export
 
 
@@ -119,57 +118,6 @@ class _TFRecordDataset(dataset_ops.DatasetSource):
     return structure.TensorStructure(dtypes.string, [])
 
 
-class ParallelInterleaveDataset(dataset_ops.UnaryDataset):
-  """A `Dataset` that maps a function over its input and flattens the result."""
-
-  def __init__(self, input_dataset, map_func, cycle_length, block_length,
-               sloppy, buffer_output_elements, prefetch_input_elements):
-    """See `tf.data.experimental.parallel_interleave()` for details."""
-    self._input_dataset = input_dataset
-    self._map_func = dataset_ops.StructuredFunctionWrapper(
-        map_func, self._transformation_name(), dataset=input_dataset)
-    if not isinstance(self._map_func.output_structure,
-                      dataset_ops.DatasetStructure):
-      raise TypeError("`map_func` must return a `Dataset` object.")
-    self._structure = self._map_func.output_structure._element_structure  # pylint: disable=protected-access
-    self._cycle_length = ops.convert_to_tensor(
-        cycle_length, dtype=dtypes.int64, name="cycle_length")
-    self._block_length = ops.convert_to_tensor(
-        block_length, dtype=dtypes.int64, name="block_length")
-    self._sloppy = ops.convert_to_tensor(
-        sloppy, dtype=dtypes.bool, name="sloppy")
-    self._buffer_output_elements = convert.optional_param_to_tensor(
-        "buffer_output_elements",
-        buffer_output_elements,
-        argument_default=2 * block_length)
-    self._prefetch_input_elements = convert.optional_param_to_tensor(
-        "prefetch_input_elements",
-        prefetch_input_elements,
-        argument_default=2 * cycle_length)
-    variant_tensor = ged_ops.experimental_parallel_interleave_dataset(
-        self._input_dataset._variant_tensor,  # pylint: disable=protected-access
-        self._map_func.function.captured_inputs,
-        self._cycle_length,
-        self._block_length,
-        self._sloppy,
-        self._buffer_output_elements,
-        self._prefetch_input_elements,
-        f=self._map_func.function,
-        **dataset_ops.flat_structure(self))
-    super(ParallelInterleaveDataset, self).__init__(input_dataset,
-                                                    variant_tensor)
-
-  def _functions(self):
-    return [self._map_func]
-
-  @property
-  def _element_structure(self):
-    return self._structure
-
-  def _transformation_name(self):
-    return "tf.data.experimental.parallel_interleave()"
-
-
 @tf_export("data.TFRecordDataset", v1=[])
 class TFRecordDatasetV2(dataset_ops.DatasetV2):
   """A `Dataset` comprising records from one or more TFRecord files."""
@@ -221,10 +169,10 @@ class TFRecordDatasetV2(dataset_ops.DatasetV2):
     if num_parallel_reads is None:
       self._impl = filenames.flat_map(read_one_file)
     else:
-      self._impl = ParallelInterleaveDataset(
-          filenames, read_one_file, cycle_length=num_parallel_reads,
-          block_length=1, sloppy=False, buffer_output_elements=None,
-          prefetch_input_elements=None)
+      self._impl = filenames.interleave(
+          read_one_file,
+          cycle_length=num_parallel_reads,
+          num_parallel_calls=num_parallel_reads)
     variant_tensor = self._impl._variant_tensor  # pylint: disable=protected-access
     super(TFRecordDatasetV2, self).__init__(variant_tensor)
